actor model changed

+++ hw02_walker2d/train.py
@@ -54,9 +54,9 @@ class Actor(nn.Module):
             nn.ReLU(),
             nn.Linear(hidden_size, hidden_size),
             nn.ReLU(),
-            nn.Linear(hidden_size, hidden_size / 2),
+            nn.Linear(hidden_size, int(hidden_size / 2)),
             nn.ReLU(),
-            nn.Linear(hidden_size / 2, 2 * action_dim),
+            nn.Linear(int(hidden_size / 2), 2 * action_dim),
         )
         self.sigma = None
